backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0